schedule validation in microtasks
