+++ src/types.ts
@@ -14,6 +14,46 @@ export interface Errors {
 
 export type ValidationTuple = [string, (value: any) => string | undefined];
 
+export interface SuccessBag {
+  resetForm: () => void;
+}
+
+export interface ErrorBag {
+  setErrors: (errors: Errors) => void;
+  setFormError: (error: string) => void;
+}
+
+export interface CallBag {
+  props?: object;
+  setErrors: (errors: Errors) => void;
+  setFormError: (error: string) => void;
+}
+
+export interface Payload {
+  change: (fieldId: string, value: any) => void;
+  formError?: string | null;
+  isDirty?: boolean | null;
+  isSubmitting?: boolean | null;
+  handleSubmit: (e?: React.SyntheticEvent) => void;
+  resetForm: () => void;
+}
+
+export interface FormOptions<T>
+  extends Omit<React.HTMLProps<HTMLFormElement>, 'onSubmit' | 'onError'> {
+  children?: ((form: Payload) => React.ReactNode) | React.ReactNode;
+  enableReinitialize?: boolean;
+  initialErrors?: Errors;
+  initialValues?: Partial<T>;
+  noForm?: boolean;
+  onError?: (error: object, callbag: ErrorBag) => void;
+  onSuccess?: (result: any, callbag: SuccessBag) => void;
+  onSubmit: (values: Partial<T>, callbag: CallBag) => Promise<any> | any;
+  shouldSubmitWhenInvalid?: boolean;
+  validate?: (values: Partial<T>) => object | undefined;
+  validateOnBlur?: boolean;
+  validateOnChange?: boolean;
+}
+
 export interface FormHookContext {
   errors: Errors;
   isDirty?: boolean;
